+++ cfgFactionClasses.hpp
@@ -1,0 +1,17 @@
+class CfgFactionClasses
+{
+	class 121st_MSV
+	{
+		displayName="121st Motostrelki Division";
+		priority=5;
+		side=1;
+		icon= "";
+	};
+	class 75th_RR
+	{
+		displayName="75th Ranger Regiment";
+		priority=5;
+		side=1;
+		icon= "";
+	};
+};
