+++ cfgPatches.hpp
@@ -18,26 +18,9 @@ class CfgPatches
 	{
 		units[]=
 		{
-			"75th_RR_Soldier_base_f", "75th_RR_Platoon_Leader", "75th_Medic", "75th_RR_Platoon_Sgt", "75th_RR_JTAC", "75th_RR_Squad_Leader", "75th_RR_FTL", "75th_RR_Rifleman", "75th_RR_AR", "75th_RR_Grenadier", "75th_RR_MG", "75th_RR_AMG", "75th_RR_AT", "75th_RR_AAT"
+			"75th_RR_Soldier_base_f", "75th_RR_Platoon_Leader", "75th_RR_Medic", "75th_RR_Platoon_Sgt", "75th_RR_JTAC", "75th_RR_Squad_Leader", "75th_RR_FTL", "75th_RR_Rifleman", "75th_RR_AR", "75th_RR_Grenadier", "75th_RR_MG", "75th_RR_AMG", "75th_RR_AT", "75th_RR_AAT"
 		};
 		requiredVersion=0.1;
 		requiredAddons[]={A3_Characters_F"};
-	};
-};
-class CfgFactionClasses
-{
-	class 121st_MSV
-	{
-		displayName="121st Motostrelki Division";
-		priority=5;
-		side=1;
-		icon= "";
-	};
-	class 75th_RR
-	{
-		displayName="75th Ranger Regiment";
-		priority=4;
-		side=1;
-		icon= "";
 	};
 };
