+++ config.cpp
@@ -2,6 +2,7 @@
 //Github repo VALHALLA
 
 	#include "cfgPatches.hpp"
+	#include "cfgFactionClasses.hpp"
 	#include "cfgWeapons_MSV.hpp"
 	#include "cfgVehicles_MSV.hpp"
 	#include "cfgUnits_MSV.hpp"
